Redirect to team events on click in calendar

+++ static/scripts/calendar.js
@@ -111,19 +111,26 @@ $(document).ready(function () {
                     action: '/calendar/events/' + event.attributes.uid
                 });
 
-                transformCourseOrTeamEvent($editEventModal, event);
-
-                $editEventModal.find('.btn-delete').click(e => {
-                    $.ajax({
-                        url: '/calendar/events/' + event.attributes.uid,
-                        type: 'DELETE',
-                        error: showAJAXError,
-                        success: function(result) {
-                            reloadCalendar();
-                        },
+                if (event["x-sc-courseId"]) { // course event
+                    transformCourseOrTeamEvent($editEventModal, event);
+                    $editEventModal.find('.btn-delete').click(e => {
+                        $.ajax({
+                            url: '/calendar/events/' + event.attributes.uid,
+                            type: 'DELETE',
+                            error: showAJAXError,
+                            success: function(result) {
+                                reloadCalendar();
+                            },
+                        });
                     });
-                });
-                $editEventModal.appendTo('body').modal('show');
+                    $editEventModal.appendTo('body').modal('show');
+                }
+
+                if (event["x-sc-teamId"]) { // team event
+                    const teamId = event["x-sc-teamId"];
+                    window.location.assign(`/teams/${teamId}?activeTab=events`);
+                }
+
             }
         },
         dayClick: function(date, jsEvent, view) {
